Bug Fixes + Experimental Test of the Calculated Model
Downloaded a copy of the notebook from github and found some bugs when you run things for the first time.
These mostly have to do with adding the radius R as an argument to some of the functions. 
I also got the idea of adding an experiment involving a household cylinder and a pitcher of water that may be fun for students.